agentura-commons/pyproject.toml
[project]
name = "agentura-commons"
version = "0.1.0"
description = "Shared MCP + A2A base classes for Semos Agentura agents"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.9",
    "a2a-sdk>=0.3",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic-settings>=2.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentura_commons"]
